refactor missionPage to index.js
