+++ src/Components/MyProfile.js
@@ -1,8 +1,10 @@
 import React from 'react';
+import MissionProfile from './missionpage/mission_myprofile';
 
 const MyProfile = () => (
   <div>
     <h1>my profile</h1>
+    <MissionProfile />
   </div>
 );
