+++ src/Redux/configureStore.js
@@ -1,15 +1,10 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
-import rocketsReducer from './rockets/rockets';
-import missionsReducer from './missions/mission';
-import dragonsReducer from './dragons/dragons';
+import missionReducer from './missions/mission';
 
 const reducer = combineReducers({
-
-  rockets: rocketsReducer,
-  missions: missionsReducer,
-  dragons: dragonsReducer,
+  missions: missionReducer,
 
 });
